refactor(api-client): split client into endpoint modules
  Extract api-client request/core utilities and move auth/import calls into
  dedicated modules. Also standardize test locations to `__tests__` and move
  web e2e specs to repo-level `tests/e2e` with updated config paths.

  Files:
  - apps/web/__tests__/integration/action-contract.test.ts
  - apps/web/playwright.config.ts
  - apps/web/vitest.config.ts
  - packages/api-client/__tests__/index.test.ts
  - packages/api-client/auth/login.ts
  - packages/api-client/auth/logout.ts
  - packages/api-client/auth/me.ts
  - packages/api-client/auth/register.ts
  - packages/api-client/core/client-options.ts
  - packages/api-client/core/parse-json.ts
  - packages/api-client/core/request.ts
  - packages/api-client/imports/get-import-run-issues.ts
  - packages/api-client/imports/get-import-run.ts
  - packages/api-client/imports/get-import-status.ts
  - packages/api-client/index.ts
  - packages/auth/__tests__/index.test.ts
  - packages/server/auth/__tests__/service.test.ts
  - tests/e2e/smoke.spec.ts
