+++ apps/web/playwright.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "@playwright/test";
 
 export default defineConfig({
-  testDir: "./tests/e2e",
+  testDir: "../../tests/e2e",
   use: {
     baseURL: "http://localhost:3000",
   },
