+++ packages/api-client/__tests__/index.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { createApiClient } from "./index";
+import { createApiClient } from "../index";
 
 describe("api client request headers", () => {
   afterEach(() => {
