+++ packages/api-client/auth/logout.ts
@@ -1,0 +1,6 @@
+import type { LogoutResponse } from "@beagle/contracts";
+import type { RequestFn } from "../core/request";
+
+export function logout(request: RequestFn) {
+  return request<LogoutResponse>("/api/auth/logout", { method: "POST" });
+}
