+++ packages/api-client/core/client-options.ts
@@ -1,0 +1,5 @@
+export type ClientOptions = {
+  baseUrl?: string;
+  credentials?: RequestCredentials;
+  headers?: HeadersInit;
+};
