+++ tests/e2e/smoke.spec.ts
